requirements.txt
django
google-generativeai
pypdf2
gunicorn
whitenoise